[refactor] OnsetMatchResult를 공통 객체로 분리
